+++ packages/cashscript/test/fixture/PriceOracle.ts
@@ -1,9 +1,9 @@
-import { padMinimallyEncodedVmNumber, flattenBinArray, secp256k1 } from '@bitauth/libauth';
+import { padMinimallyEncodedVmNumber, flattenBinArray } from '@bitauth/libauth';
 import { encodeInt, sha256 } from '@cashscript/utils';
-import { SignatureAlgorithm } from '../../src/index.js';
+import { SignatureAlgorithm, SignatureTemplate } from '../../src/index.js';
 
 export class PriceOracle {
-  constructor(public privateKey: Uint8Array) {}
+  constructor(public privateKey: Uint8Array) { }
 
   // Encode a blockHeight and bchUsdPrice into a byte sequence of 8 bytes (4 bytes per value)
   createMessage(blockHeight: bigint, bchUsdPrice: bigint): Uint8Array {
@@ -14,11 +14,7 @@ export class PriceOracle {
   }
 
   signMessage(message: Uint8Array, signatureAlgorithm: SignatureAlgorithm = SignatureAlgorithm.SCHNORR): Uint8Array {
-    const signature = signatureAlgorithm === SignatureAlgorithm.SCHNORR ?
-      secp256k1.signMessageHashSchnorr(this.privateKey, sha256(message)) :
-      secp256k1.signMessageHashDER(this.privateKey, sha256(message));
-
-    if (typeof signature === 'string') throw new Error();
-    return signature;
+    const signatureTemplate = new SignatureTemplate(this.privateKey, undefined, signatureAlgorithm);
+    return signatureTemplate.signMessageHash(sha256(message));
   }
 }